+++ src/components/common/Icon.tsx
@@ -5,62 +5,62 @@ import { StyleProp, TextStyle, ViewStyle } from 'react-native';
 import IonIcon from 'react-native-vector-icons/Ionicons';
 import { IconProps } from 'react-native-vector-icons/Icon';
 
-import ArrowBack from '../assets/images/icons/arrow-back.svg';
-import ArrowDown from '../assets/images/icons/arrow-down.svg';
-import ArrowForward from '../assets/images/icons/arrow-forward.svg';
-import ArrowLeft from '../assets/images/icons/arrow-left.svg';
-import ArrowRight from '../assets/images/icons/arrow-right.svg';
-import ArrowUp from '../assets/images/icons/arrow-up.svg';
-import Close from '../assets/images/icons/close.svg';
-import Info from '../assets/images/icons/info.svg';
-import Layers from '../assets/images/icons/layers.svg';
-import Locate from '../assets/images/icons/locate.svg';
-import MapMarker from '../assets/images/icons/map-marker.svg';
-import Map from '../assets/images/icons/map.svg';
-import Menu from '../assets/images/icons/menu.svg';
-import Mic from '../assets/images/icons/mic.svg';
-import Minus from '../assets/images/icons/minus.svg';
-import OpenInNew from '../assets/images/icons/open-in-new.svg';
-import Pause from '../assets/images/icons/pause.svg';
-import Play from '../assets/images/icons/play.svg';
-import Plus from '../assets/images/icons/plus.svg';
-import RadioButtonOff from '../assets/images/icons/radio-button-off.svg';
-import RadioButtonOn from '../assets/images/icons/radio-button-on.svg';
-import RainDark from '../assets/images/icons/rain-dark.svg';
-import RainLight from '../assets/images/icons/rain-light.svg';
-import Search from '../assets/images/icons/search.svg';
-import StarSelected from '../assets/images/icons/star-selected.svg';
-import StarUnselected from '../assets/images/icons/star-unselected.svg';
-import Sunrise from '../assets/images/icons/sunrise.svg';
-import WarningsFloodLevel2 from '../assets/images/icons/warnings-flood-level-2.svg';
-import WarningsFloodLevel3 from '../assets/images/icons/warnings-flood-level-3.svg';
-import WarningsFloodLevel4 from '../assets/images/icons/warnings-flood-level-4.svg';
-import WarningsForestFireWeatherOrange from '../assets/images/icons/warnings-forest-fire-weather-orange.svg';
-import WarningsForestFireWeatherRed from '../assets/images/icons/warnings-forest-fire-weather-red.svg';
-import WarningsForestFireWeatherYellow from '../assets/images/icons/warnings-forest-fire-weather-yellow.svg';
-import WarningsGrassFireWeather from '../assets/images/icons/warnings-grass-fire-weather.svg';
-import WarningsHotWeatherOrange from '../assets/images/icons/warnings-hot-weather-orange.svg';
-import WarningsHotWeatherRed from '../assets/images/icons/warnings-hot-weather-red.svg';
-import WarningsHotWeatherYellow from '../assets/images/icons/warnings-hot-weather-yellow.svg';
-import WarningsPedestrianSafety from '../assets/images/icons/warnings-pedestrian-safety.svg';
-import WarningsRainOrange from '../assets/images/icons/warnings-rain-orange.svg';
-import WarningsRainRed from '../assets/images/icons/warnings-rain-red.svg';
-import WarningsRainYellow from '../assets/images/icons/warnings-rain-yellow.svg';
-import WarningsStatusOrange from '../assets/images/icons/warnings-status-orange.svg';
-import WarningsThunderStormOrange from '../assets/images/icons/warnings-thunder-storm-orange.svg';
-import WarningsThunderStormRed from '../assets/images/icons/warnings-thunder-storm-red.svg';
-import WarningsThunderStormYellow from '../assets/images/icons/warnings-thunder-storm-yellow.svg';
-import WarningsTrafficWeatherOrange from '../assets/images/icons/warnings-traffic-weather-orange.svg';
-import WarningsTrafficWeatherYellow from '../assets/images/icons/warnings-traffic-weather-yellow.svg';
-import WarningsUvNote from '../assets/images/icons/warnings-uv-note.svg';
-import WarningsWindOrange from '../assets/images/icons/warnings-wind-orange.svg';
-import WarningsWindRed from '../assets/images/icons/warnings-wind-red.svg';
-import WarningsWindYellow from '../assets/images/icons/warnings-wind-yellow.svg';
-import Warnings from '../assets/images/icons/warnings.svg';
-import Weather from '../assets/images/icons/weather.svg';
-import WindArrow from '../assets/images/icons/wind-arrow.svg';
-import WindDark from '../assets/images/icons/wind-dark.svg';
-import WindLight from '../assets/images/icons/wind-light.svg';
+import ArrowBack from '../../assets/images/icons/arrow-back.svg';
+import ArrowDown from '../../assets/images/icons/arrow-down.svg';
+import ArrowForward from '../../assets/images/icons/arrow-forward.svg';
+import ArrowLeft from '../../assets/images/icons/arrow-left.svg';
+import ArrowRight from '../../assets/images/icons/arrow-right.svg';
+import ArrowUp from '../../assets/images/icons/arrow-up.svg';
+import Close from '../../assets/images/icons/close.svg';
+import Info from '../../assets/images/icons/info.svg';
+import Layers from '../../assets/images/icons/layers.svg';
+import Locate from '../../assets/images/icons/locate.svg';
+import MapMarker from '../../assets/images/icons/map-marker.svg';
+import Map from '../../assets/images/icons/map.svg';
+import Menu from '../../assets/images/icons/menu.svg';
+import Mic from '../../assets/images/icons/mic.svg';
+import Minus from '../../assets/images/icons/minus.svg';
+import OpenInNew from '../../assets/images/icons/open-in-new.svg';
+import Pause from '../../assets/images/icons/pause.svg';
+import Play from '../../assets/images/icons/play.svg';
+import Plus from '../../assets/images/icons/plus.svg';
+import RadioButtonOff from '../../assets/images/icons/radio-button-off.svg';
+import RadioButtonOn from '../../assets/images/icons/radio-button-on.svg';
+import RainDark from '../../assets/images/icons/rain-dark.svg';
+import RainLight from '../../assets/images/icons/rain-light.svg';
+import Search from '../../assets/images/icons/search.svg';
+import StarSelected from '../../assets/images/icons/star-selected.svg';
+import StarUnselected from '../../assets/images/icons/star-unselected.svg';
+import Sunrise from '../../assets/images/icons/sunrise.svg';
+import WarningsFloodLevel2 from '../../assets/images/icons/warnings-flood-level-2.svg';
+import WarningsFloodLevel3 from '../../assets/images/icons/warnings-flood-level-3.svg';
+import WarningsFloodLevel4 from '../../assets/images/icons/warnings-flood-level-4.svg';
+import WarningsForestFireWeatherOrange from '../../assets/images/icons/warnings-forest-fire-weather-orange.svg';
+import WarningsForestFireWeatherRed from '../../assets/images/icons/warnings-forest-fire-weather-red.svg';
+import WarningsForestFireWeatherYellow from '../../assets/images/icons/warnings-forest-fire-weather-yellow.svg';
+import WarningsGrassFireWeather from '../../assets/images/icons/warnings-grass-fire-weather.svg';
+import WarningsHotWeatherOrange from '../../assets/images/icons/warnings-hot-weather-orange.svg';
+import WarningsHotWeatherRed from '../../assets/images/icons/warnings-hot-weather-red.svg';
+import WarningsHotWeatherYellow from '../../assets/images/icons/warnings-hot-weather-yellow.svg';
+import WarningsPedestrianSafety from '../../assets/images/icons/warnings-pedestrian-safety.svg';
+import WarningsRainOrange from '../../assets/images/icons/warnings-rain-orange.svg';
+import WarningsRainRed from '../../assets/images/icons/warnings-rain-red.svg';
+import WarningsRainYellow from '../../assets/images/icons/warnings-rain-yellow.svg';
+import WarningsStatusOrange from '../../assets/images/icons/warnings-status-orange.svg';
+import WarningsThunderStormOrange from '../../assets/images/icons/warnings-thunder-storm-orange.svg';
+import WarningsThunderStormRed from '../../assets/images/icons/warnings-thunder-storm-red.svg';
+import WarningsThunderStormYellow from '../../assets/images/icons/warnings-thunder-storm-yellow.svg';
+import WarningsTrafficWeatherOrange from '../../assets/images/icons/warnings-traffic-weather-orange.svg';
+import WarningsTrafficWeatherYellow from '../../assets/images/icons/warnings-traffic-weather-yellow.svg';
+import WarningsUvNote from '../../assets/images/icons/warnings-uv-note.svg';
+import WarningsWindOrange from '../../assets/images/icons/warnings-wind-orange.svg';
+import WarningsWindRed from '../../assets/images/icons/warnings-wind-red.svg';
+import WarningsWindYellow from '../../assets/images/icons/warnings-wind-yellow.svg';
+import Warnings from '../../assets/images/icons/warnings.svg';
+import Weather from '../../assets/images/icons/weather.svg';
+import WindArrow from '../../assets/images/icons/wind-arrow.svg';
+import WindDark from '../../assets/images/icons/wind-dark.svg';
+import WindLight from '../../assets/images/icons/wind-light.svg';
 
 type CustomIconProps = IconProps & {
   name: string;
