+++ src/components/weather/charts/TemperatureLineChart.tsx
@@ -2,42 +2,35 @@ import React from 'react';
 import { View, StyleSheet, Text } from 'react-native';
 import {
   VictoryChart,
-  VictoryAxis,
   VictoryLine,
-  VictoryArea,
+  VictoryAxis,
   VictoryLabel,
 } from 'victory-native';
 import { useTheme } from '@react-navigation/native';
 import { useTranslation } from 'react-i18next';
 
-import Icon from './Icon';
+import { CustomTheme } from '../../../utils/colors';
+import { TimestepData } from '../../../store/forecast/types';
+import chartTheme from '../../../utils/chartTheme';
 
-import { CustomTheme } from '../utils/colors';
-import chartTheme from '../utils/chartTheme';
-import { TimestepData } from '../store/forecast/types';
-
-type WindLineChartProps = {
+type TemperatureLineChartProps = {
   data: TimestepData[];
 };
 
-const WindLineChart: React.FC<WindLineChartProps> = ({ data }) => {
+const TemperatureLineChart: React.FC<TemperatureLineChartProps> = ({
+  data,
+}) => {
   const { colors } = useTheme() as CustomTheme;
   const { t } = useTranslation();
 
-  const windData = data.map((hour) => ({
+  const temperatureData = data.map((hour) => ({
     x: hour.epochtime,
-    y: hour.windspeedms,
+    y: hour.temperature,
   }));
 
-  const gustData = data.map((hour) => ({
+  const feelsLikeData = data.map((hour) => ({
     x: hour.epochtime,
-    y: hour.hourlymaximumgust,
-  }));
-
-  const combinedData = data.map((hour) => ({
-    x: hour.epochtime,
-    y0: hour.windspeedms,
-    y: hour.hourlymaximumgust,
+    y: hour.feelsLike,
   }));
 
   const hourGetter = (datum: any): string => {
@@ -45,36 +38,20 @@ const WindLineChart: React.FC<WindLineChartProps> = ({ data }) => {
     return `${date.getHours()}`;
   };
 
-  const max = Math.max(...windData.map((d) => d.y));
+  const max = Math.max(
+    ...temperatureData.concat(feelsLikeData).map((d) => d.y)
+  );
+  const min = Math.min(
+    ...temperatureData.concat(feelsLikeData).map((d) => d.y)
+  );
 
-  const WindLabel = (asd: any) => {
-    const index = Number(asd.index);
-    const windDir = data[index]?.winddirection;
-    return (
-      <View style={[styles.arrowStyle, { left: asd.x - 10 }]}>
-        <Icon
-          name="wind-arrow"
-          width={20}
-          height={20}
-          style={{
-            color: colors.primaryText,
-            transform: [
-              {
-                rotate: `${windDir + 45 - 180}deg`,
-              },
-            ],
-          }}
-        />
-      </View>
-    );
-  };
+  // TODO: refine logic
+  const domainMax = max >= 25 ? max + 5 : 25;
+  const domainMin = min <= -5 ? min - 5 : -5;
 
   return (
     <View style={styles.container}>
-      <VictoryChart
-        height={300}
-        theme={chartTheme}
-        padding={{ top: 50, right: 50, left: 50, bottom: 35 }}>
+      <VictoryChart height={300} theme={chartTheme}>
         <VictoryAxis
           fixLabelOverlap
           style={{
@@ -83,47 +60,39 @@ const WindLineChart: React.FC<WindLineChartProps> = ({ data }) => {
                 Number(index) === 0 ? 'none' : colors.primaryText,
             },
           }}
+          offsetY={50}
         />
         <VictoryAxis
           dependentAxis
-          domain={[0, max + 2]}
+          crossAxis={false}
+          domain={[domainMin, domainMax]}
           style={{
             tickLabels: {
               fill: colors.primaryText,
             },
           }}
         />
-        <VictoryArea
-          data={combinedData}
-          x={hourGetter}
-          style={{ data: { fill: '#d8d8d8' } }}
-          animate={{ duration: 500, onLoad: { duration: 250 } }}
-        />
         <VictoryLine
-          data={windData}
-          labels={({ datum }) => `${datum}`}
-          labelComponent={<WindLabel />}
-          domain={{ y: [0, 8] }}
+          data={temperatureData}
+          domain={{ y: [0, 25] }}
           animate={{ duration: 500, onLoad: { duration: 250 } }}
           style={{ data: { stroke: colors.primaryText } }}
           x={hourGetter}
           interpolation="natural"
         />
+
         <VictoryLine
-          data={gustData}
+          data={feelsLikeData}
           domain={{ y: [0, 25] }}
           animate={{ duration: 500, onLoad: { duration: 250 } }}
           style={{
-            data: {
-              stroke: colors.chartSecondaryLine,
-              strokeDasharray: '4',
-            },
+            data: { stroke: colors.chartSecondaryLine, strokeDasharray: '4' },
           }}
           x={hourGetter}
           interpolation="natural"
         />
         <VictoryLabel
-          text="m/s"
+          text="°C"
           x={30}
           y={20}
           style={{ fill: colors.primaryText }}
@@ -140,7 +109,7 @@ const WindLineChart: React.FC<WindLineChartProps> = ({ data }) => {
             ]}
           />
           <Text style={[styles.legendText, { color: colors.primaryText }]}>
-            {t('forecast:charts:windSpeed')} (m/s)
+            {t('forecast:charts:temperature')} (°C)
           </Text>
         </View>
         <View style={styles.legendRow}>
@@ -167,23 +136,7 @@ const WindLineChart: React.FC<WindLineChartProps> = ({ data }) => {
             />
           </View>
           <Text style={[styles.legendText, { color: colors.primaryText }]}>
-            {t('forecast:charts:windGust')} (m/s)
-          </Text>
-        </View>
-        <View style={styles.legendRow}>
-          <Icon
-            name="wind-arrow"
-            width={20}
-            height={20}
-            style={[
-              styles.iconMargin,
-              {
-                color: colors.primaryText,
-              },
-            ]}
-          />
-          <Text style={[styles.legendText, { color: colors.primaryText }]}>
-            {t('forecast:charts:windDirection')}
+            {t('forecast:charts:feelsLike')} (°C)
           </Text>
         </View>
       </View>
@@ -196,10 +149,6 @@ const styles = StyleSheet.create({
     flex: 1,
     justifyContent: 'center',
     alignItems: 'center',
-  },
-  arrowStyle: {
-    position: 'absolute',
-    top: 25,
   },
   legendContainer: {
     flex: 1,
@@ -226,9 +175,6 @@ const styles = StyleSheet.create({
   withMarginRight: {
     marginRight: 2,
   },
-  iconMargin: {
-    marginLeft: -4,
-  },
 });
 
-export default WindLineChart;
+export default TemperatureLineChart;
