+++ src/components/weather/wrappers/DaySelectorWrapper.tsx
@@ -3,9 +3,9 @@ import { View, StyleSheet, TouchableOpacity, Text } from 'react-native';
 import { useTheme } from '@react-navigation/native';
 import { useTranslation } from 'react-i18next';
 
-import Icon from './Icon';
+import Icon from '../../common/Icon';
 
-import { CustomTheme } from '../utils/colors';
+import { CustomTheme } from '../../../utils/colors';
 
 type DaySelectorWrapperProps = {
   selectedDate: string | false | undefined;
